# detect: initialization optimization