A lot of time was spent in `SigMatchListSMBelongsTo` for the `mpm_sm`.

Optimize this by keeping the value at hand during Signature parsing and
detection engine setup.


Results on my machine loading:

ET Open:

- master: 28s
- this branch: 13s

ET Pro:

- master: 109s
- this branch: 45s