@suricata-qa 2022-04-01
Information: QA ran without warnings. 

Pipeline 6711